Support for `g:fzf_action` to change hotkeys.
With fzf, if you specify the following (which is the default):
```
let g:fzf_action = {
  \ 'ctrl-t': 'tab split',
  \ 'ctrl-x': 'split',
  \ 'ctrl-v': 'vsplit' }
```

It is possible to override the default hotkey for opening files from the fzf view using `g:fzf_action`.
In my case i have `ctl-x` already mapped to an action, so i cannot open
a file in split-view using fzf/skim. Changing the `g:fzf_action` allows
me to change the hotkey for split-view to `ctrl-e` (or anything)

```
let g:fzf_action = {
  \ 'ctrl-t': 'tab split',
  \ 'ctrl-e': 'split',
  \ 'ctrl-v': 'vsplit' }
```