fix: check version_requirement components numerically
The input parameters are strings that get split into string components.
When such strings are compared with <> they will perform a string
compare instead of numerical compare which means 0.10 and 0.9 will not
be compared as expected but by their individual chars. This leads to 1
getting compared against 9 leading to an unsuccessful check.

By using str2nr we create local numeric variables of the current
components and therefore perform a numeric compare of 10 against 9
leading to a successful check.

Signed-off-by: Levente Polyak <levente@leventepolyak.net>