fix: support future clap-styled version output of skim
The clap crate >= 3 as used in skim has a new default version option
that resolves to a different output style including the application name
in front of the actual version.

Note that a match at the start is preferred, thus when the pattern is
using "*" (any number of matches) it tends to find zero matches at the
start instead of a number of matches further down in the text.

Hence change the matcher to \+ so it requires >= 1 characters from the
range to match, which avoids matching against nothing at the start.
This allows to have a future proof version matcher which is also
backwards compatible.

Signed-off-by: Levente Polyak <levente@leventepolyak.net>